==> Scripts/S1_UMI_BC_include_more_cells.Rmd <==

---
title: "VDJ sequencing - HUSTLE-Seq - UMI counts + cell barcode exctraction"
output:
  html_document:
    df_print: paged
  pdf_document: default
  toc: TRUE
---


```{r load_libraries, echo = F, message = F, warning = F}
library(Seurat)
library(patchwork)
library(readxl)
library(tidyverse)
library(ggplot2)
library(ggpubr)
library(reshape2)
library(broom)
library(stringr)
library(ggplot2)
library(cowplot)
library(RColorBrewer)
library(data.table)
library(DropletUtils)
```

```{r datadiscovery, echo = F}

samplesheet <- read_excel("../Data/a_2024_HUSTLEseq_scRNAseq_nucleotide_sequencing_round2.xlsx")

colnames(samplesheet) <- sapply(c(1:ncol(samplesheet)), function(x) paste0(samplesheet[1,x], colnames(samplesheet)[x], collapse = ""))

samplesheet <- samplesheet %>% 
  filter(!row_number() %in% c(1)) %>% 
  dplyr::rename(Library_name = `NameLibrary Name`) %>% 
  dplyr::rename(Antibody = `Multiplex Source Tag...31`)

lib_name = "D13_D33_D45_D47_CD4_Naive"
d <- samplesheet %>% filter(Library_name == lib_name)


# Antibody column needs to be changed from Total_Seq to TotalSeq
samplesheet <- samplesheet %>%
  mutate(Antibody = gsub("Total_Seq", "TotalSeq", Antibody))
```

The hastag antibodies used were:
```{r hto_used}

# lib name
hto_used = samplesheet %>%
    filter(Library_name == lib_name) %>%
    pull(Antibody) %>%
    unique

hto_used
```

Load raw features
```{r load_libs}
# Load the PBMC dataset
data <- Read10X(data.dir = paste0("../Data/",lib_name,"/raw_feature_bc_matrix/"))

gex <- data[[1]]
hto <- data[["Antibody Capture"]][rownames(data[["Antibody Capture"]]) %in% hto_used, ]

options(Seurat.object.assay.calcn = TRUE)
hustle_raw <- CreateSeuratObject(counts = gex, project = lib_name, min.cells = 3, min.features = 200)
hustle_raw
```


```{r}
# RNA cells in Seurat object
rna_cells <- colnames(hustle_raw)

# HTO barcodes
hto_cells <- colnames(hto)

# Overlap?
length(intersect(rna_cells, hto_cells))

common_cells <- intersect(colnames(hustle_raw), colnames(hto))
hto_filtered <- hto[, common_cells]

# Make sure order matches
hto_filtered <- hto_filtered[, colnames(hustle_raw)]

# Now it will work
hustle_raw[["HTO"]] <- CreateAssayObject(counts = hto_filtered)

```

Load filtered features
```{r load_libs}
# Load the PBMC dataset
data <- Read10X(data.dir = paste0("../Data/",lib_name,"/filtered_feature_bc_matrix/"))

gex <- data[[1]]
hto <- data[["Antibody Capture"]][rownames(data[["Antibody Capture"]]) %in% hto_used, ]

options(Seurat.object.assay.calcn = TRUE)
hustle_filtered <- CreateSeuratObject(counts = gex, project = lib_name, min.cells = 3, min.features = 200)
hustle_filtered
```


```{r}
# make UMI counts vs barcodes plot

# Extract UMI counts - raw
umi_counts_raw <- hustle_raw@meta.data$nCount_RNA
umi_counts_raw <- sort(umi_counts_raw, decreasing = TRUE)  # Sort in descending order

# Prepare data for plotting
umi_data_raw <- data.frame(
  Rank = 1:length(umi_counts_raw),
  UMI = umi_counts_raw
)

threshold <- 700  # Example threshold
p1 <- ggplot(umi_data_raw, aes(x = Rank, y = UMI)) +
  geom_point(size = 0.5) +
  geom_hline(yintercept = threshold, linetype = "dashed", color = "red") +
  annotate(
    "text", 
    x = max(umi_data_raw$Rank) * 0.8,  # Adjust x position (80% of the x-axis range)
    y = threshold, 
    label = paste("Threshold =", threshold),
    color = "red",
    hjust = 0,  # Horizontal alignment of the text
    vjust = -0.5  # Vertical alignment of the text (above the line)
  ) +
  scale_y_continuous( limits=c(0,max(umi_data_raw$UMI)), 
    breaks = c(1, 10, 100, 1000, 10000),
                     labels = c("1", "10", "100", "1000", "10000")) +
  scale_y_log10(
   # breaks = c(1, 10, 100, 1000, 10000),  # Specify desired ticks
    #labels = c("1", "10", "100", "1000", "10000")
  ) +
  scale_x_log10() +
  labs(
    x = "Barcodes (ranked)",
    y = "UMI Counts",
    title = "UMI Counts vs. Barcodes raw features"
  ) +
  theme_minimal()

## filtered features

# Extract UMI counts
umi_counts_filtered <- hustle_filtered@meta.data$nCount_RNA
umi_counts_filtered <- sort(umi_counts_filtered, decreasing = TRUE)  # Sort in descending order

# Prepare data for plotting
umi_data_filtered <- data.frame(
  Rank = 1:length(umi_counts_filtered),
  UMI = umi_counts_filtered
)

p2 <- ggplot(umi_data_filtered, aes(x = Rank, y = UMI)) +
  geom_point(size = 0.5) +
  geom_hline(yintercept = threshold, linetype = "dashed", color = "red") +
  annotate(
    "text", 
    x = max(umi_data_filtered$Rank) * 0.8,  # Adjust x position (80% of the x-axis range)
    y = threshold, 
    label = paste("Threshold =", threshold),
    color = "red",
    hjust = 0,  # Horizontal alignment of the text
    vjust = -0.5  # Vertical alignment of the text (above the line)
  ) +
  scale_y_continuous( limits=c(0,max(umi_data_filtered$UMI)), 
    breaks = c(1, 10, 100, 1000, 10000),
                     labels = c("1", "10", "100", "1000", "10000")) +
  scale_y_log10(
   # breaks = c(1, 10, 100, 1000, 10000),  # Specify desired ticks
   # labels = c("1", "10", "100", "1000", "10000")
  ) +
  scale_x_log10() +
  labs(
    x = "Barcodes (ranked)",
    y = "UMI Counts",
    title = "UMI Counts vs. Barcodes filtered features"
  ) +
  theme_minimal()

p1 + p2
#ggsave("UMI counts vs barcodes raw and filtered features.pdf")
```

```{r}
# save barcodes with UMI counts above a certain threshold
# Subset the Seurat object to retain only cells above the threshold
threshold = 700
seurat_filtered <- subset(hustle_raw, subset = nCount_RNA > threshold)

saveRDS(seurat_filtered, "robjects/D13_D33_D45_D47_CD4_Naive_filtered_17k.rds")
```

```{r}

# Display the number of cells retained
cat("Number of cells above threshold:", ncol(seurat_filtered), "\n")

# Optional: View metadata of the filtered cells
head(seurat_filtered@meta.data)

all_data <- rbind(
  data.frame(UMI = hustle_raw@meta.data$nCount_RNA, Dataset = "All Barcodes"),
  data.frame(UMI = seurat_filtered@meta.data$nCount_RNA, Dataset = "Filtered Barcodes")
)

# Create a faceted histogram
ggplot(all_data, aes(x = UMI, fill = Dataset)) +
  geom_histogram(
    bins = 50, 
    color = "black", 
    alpha = 0.7
  ) +
  scale_x_log10() +  # Optional: log10 scale for better visualization
  labs(
    title = "UMI Counts Before and After Filtering",
    x = "UMI Counts",
    y = "Frequency"
  ) +
  facet_wrap(~Dataset) +
  theme_minimal()


#ggsave("UMI counts histogram raw and filtered features.pdf")

```
