==> Scripts/S2_0_allsteps.Rmd <==
---
title: "VDJ sequencing - HUSTLE-Seq"
output:
  html_document:
    df_print: paged
    code_folding: hide
  pdf_document: default
  toc: TRUE
---

```{r opts, echo = F}
knitr::opts_chunk$set(warning = F, message = F)
# knitr::opts_chunk$set(fig.pos = "!H", out.extra = "")
#knitr::opts_chunk$set(tidy.opts=list(width.cutoff=80),tidy=TRUE)
knitr::opts_chunk$set(fig.width=12, fig.height=8) 

```

```{r load_libraries, echo = F, message = F, warning = F}
library(Seurat)
library(patchwork)
library(readxl)
library(tidyverse)
library(ggplot2)
library(ggpubr)
library(reshape2)
library(broom)
library(stringr)
library(ggplot2)
library(cowplot)
library(RColorBrewer)
library(data.table)
library(DropletUtils)
library(scCustomize)
#library(hdfr5)
#remotes::install_github('chris-mcginnis-ucsf/DoubletFinder')
#library(DoubletFinder)
# devtools::install_github("satijalab/seurat-data")
#devtools::install_github("satijalab/azimuth")
#library(Azimuth)
# library(SeuratData)
```

```{r QC, child = 'S2_1_QC_demultiplex_Naive_custom_CellSNP.Rmd', echo = T, result = 'asis'}

```

```{r process, child = 'S2_2_normalisation_DimRed_UMAP.Rmd', echo = T, result = 'asis'}

```
