==> Scripts/S4_post_integration_exploration.Rmd <==
---
title: "R Notebook"
output:
  html_document:
    df_print: paged
    code_folding: hide
  pdf_document: default
  toc: TRUE
---

```{r opts, echo = F}
knitr::opts_chunk$set(warning = F, message = F)
# knitr::opts_chunk$set(fig.pos = "!H", out.extra = "")
#knitr::opts_chunk$set(tidy.opts=list(width.cutoff=80),tidy=TRUE)
knitr::opts_chunk$set(fig.width=10, fig.height=8) 

```


```{r load_libraries, echo = F, message = F, warning = F}
library(Seurat)
library(patchwork)
library(readxl)
library(tidyverse)
library(ggplot2)
library(ggpubr)
library(reshape2)
library(broom)
library(stringr)
library(ggplot2)
library(cowplot)
library(RColorBrewer)
library(data.table)
library(DropletUtils)
# devtools::install_github("satijalab/seurat", "seurat5")
# devtools::install_github("satijalab/seurat-data", "seurat5")
# BiocManager::install("DirichletMultinomial")
# install.packages("hdf5r")
# remotes::install_github("mojaveazure/seurat-disk")
# BiocManager::install("TFBSTools")
# remotes::install_github('satijalab/azimuth', ref = 'master')
library(Azimuth)
library(plyr)

```

```{r}
integrated <- readRDS("robjects/naive_merged_with_integrated_proliferated_L1_L2_custom_naive_cellsnp.rds")
```

```{r}
table(integrated$orig.ident)

table(integrated$orig.ident, integrated$Donor_expt)
```


```{r, fig.width=10, fig.height=6}
mu_nC <- ddply(integrated@meta.data, c("orig.ident","Donor_expt"), 
               summarise, 
               grp.mean=median(nCount_RNA))

nC <- ggplot(integrated@meta.data, aes(x = nCount_RNA, fill = Donor_expt)) +
  geom_histogram(bins = 50, alpha = 0.6, position = "identity") + 
  facet_wrap(~ orig.ident) +
  theme_minimal(base_size = 10) +
  labs(
    x = "nCount_RNA",
    y = "Cell count",
    title = "nCount_RNA distribution by library"
  )

mu_nF <- ddply(integrated@meta.data, c("orig.ident","Donor_expt"), 
               summarise, 
               grp.mean=median(nFeature_RNA))

nF <- ggplot(integrated@meta.data, aes(x = nFeature_RNA, fill = Donor_expt)) +
  geom_histogram(bins = 50, alpha = 0.6, position = "identity") + 
  facet_wrap(~ orig.ident) +
  theme_minimal(base_size = 10) +
  labs(
    x = "nFeature_RNA",
    y = "Cell count",
    title = "nFeature_RNA distribution by library"
  )

mu_mt <- ddply(integrated@meta.data, c("orig.ident","Donor_expt"), 
               summarise, 
               grp.mean=median(percent.mt))

mt <- ggplot(integrated@meta.data, aes(x = percent.mt, fill = Donor_expt)) +
  geom_histogram(bins = 50, alpha = 0.6, position = "identity") + 
  facet_wrap(~ orig.ident) +
  theme_minimal(base_size = 10) +
  labs(
    x = "percent.MT",
    y = "Cell count",
    title = "percent.MT distribution by library"
  )


(nC + nF / mt)

#ggsave("QCmetric_integrated.object.png", dpi = 300, width = 25, height = 20, units = "cm")
```


```{r}
DefaultAssay(integrated) <- "RNA"
DimHeatmap(integrated, reduction = "pca", dims = 1:5, cells = 500, balanced = TRUE)
```

```{r}
DefaultAssay(integrated) <- "RNA"
FeaturePlot(integrated, features = c("XIST"))

```

<!-- ```{r} -->
<!-- DefaultAssay(integrated) <- "RNA" -->
<!-- DimPlot(integrated, group.by = "seurat_clusters",reduction = "umap_cca", label = "T") -->
<!-- #ggsave("umap_integrated_s.clusters.png", dpi = 300) -->
<!-- ``` -->

```{r}
integrated <- FindClusters(integrated, resolution = 0.5)
DimPlot(integrated, group.by = "seurat_clusters",reduction = "umap_cca", label = "T") + ggtitle("res=0.5")
```

```{r}
DimPlot(integrated, group.by = "orig.ident", ,reduction = "umap_cca", label = F, alpha = 0.7)
#ggsave("umap_integrated_orig.ident.png", dpi = 300, width = 25, units = "cm")

```

```{r}
DimPlot(integrated, group.by = "Donor_expt", ,reduction = "umap_cca", label = F)
#ggsave("umap_integrated_donor.png", dpi = 300, width = 25, units = "cm")

table(integrated$Donor_expt)
```


## Features in clusters
```{r}
myPalette <- colorRampPalette(rev(brewer.pal(11, "Spectral")))
lib_name = "integrated"

meta <- integrated@meta.data %>% 
  rownames_to_column("Cell") %>% 
  select("Cell", "nFeature_RNA", "percent.mt", "nCount_RNA")

umap_embed <- integrated@reductions[["umap_cca"]]@cell.embeddings %>% 
  as.data.frame() %>% 
  rownames_to_column("Cell") %>%
  left_join(., meta) %>% 
  mutate(lib = lib_name)

# mt.percent on umap
sc <- scale_colour_gradientn(colours = myPalette(100), limits=c(0, max(umap_embed$percent.mt)))
ggplot(umap_embed, aes(umapcca_1, umapcca_2, colour = percent.mt)) +
  geom_point(size = 0.3, alpha = 0.8) +
  facet_wrap( ~ lib, ncol = 2, scales = "free") +
  sc +
  theme_bw()

# nF on umap
sc <- scale_colour_gradientn(colours = myPalette(100), limits=c(min(umap_embed$nFeature_RNA), max(umap_embed$nFeature_RNA)))
ggplot(umap_embed, aes(umapcca_1, umapcca_2, colour = nFeature_RNA)) +
  geom_point(size = 0.3, alpha = 0.8) +
  facet_wrap( ~ lib, ncol = 2, scales = "free") +
  sc +
  theme_bw()

# nCount on umap
sc <- scale_colour_gradientn(colours = myPalette(100), limits=c(min(umap_embed$nCount_RNA), max(umap_embed$nCount_RNA)))
ggplot(umap_embed, aes(umapcca_1, umapcca_2, colour = nCount_RNA)) +
  geom_point(size = 0.3, alpha = 0.8) +
  facet_wrap( ~ lib, ncol = 2, scales = "free") +
  sc +
  theme_bw()

```

```{r}
#QC coloured by clusters

# three clusters
table(integrated$RNA_snn_res.1)
cluster.set <- unique(integrated$RNA_snn_res.1)
# overall nFeature
sum(rowSums(integrated[['RNA']]) != 0)

# nFeature for each cluster
nF <- sapply(X = cluster.set, function(c) {
  cells.c <- WhichCells(object = integrated, expression = RNA_snn_res.1 == c)
  #nFeature.c <- sum(rowSums(mono[['RNA']]@counts[, cells.c ]) != 0) / length(cells.c) # normalised by number of cells in a cluster
  df.cells <- integrated@meta.data[which(rownames(integrated@meta.data) %in% cells.c),]
  nFeature.c <- sum(df.cells$nFeature_RNA) / length(cells.c)
  return(nFeature.c)
}
)

#nC <- sapply(cluster.set, function(x) sum(x == mono$RNA_snn_res.0.9))

nC <- sapply(X = cluster.set, function(c) {
  cells.c <- WhichCells(object = integrated, expression = RNA_snn_res.1 == c)
  df.cells <- integrated@meta.data[which(rownames(integrated@meta.data) %in% cells.c),]
  nCount.c <- sum(df.cells$nCount_RNA) / length(cells.c)
  return(nCount.c)
}
)

mt <- sapply(X = cluster.set, function(c) {
  cells.c <- WhichCells(object = integrated, expression = RNA_snn_res.1 == c)
  mt.df <- integrated@meta.data[which(rownames(integrated@meta.data) %in% cells.c),]
  mt <- sum(integrated$percent.mt) / length(cells.c) # [ cells.c ] != 0, normalised by number of cells in a cluster
  return(mt)
}
)

# First the cluster annotation and the tsne embeddings are merged
label.df <- cbind(as.data.frame(integrated@meta.data), as.data.frame(integrated@reductions$umap_cca@cell.embeddings)) 

# using dplyr across to calculate the mean mitochondrial percentage and 
# the median tsne values per cluster
label.df <- label.df %>% dplyr::group_by(seurat_clusters) %>% dplyr::summarise(dplyr::across(percent.mt, ~ mean(.), .names = "mean_{.col}"), dplyr::across(contains("integratedcca"), ~ median(.)))

qc_df <- data.frame(seurat_clusters = cluster.set, nFeature_RNA = nF, nCount_RNA = nC) %>% 
  left_join(., label.df)

ggplot(qc_df, aes(x = nFeature_RNA, y = nCount_RNA, size = nCount_RNA, colour = mean_percent.mt)) + #, label = seurat_clusters
  geom_point(alpha = 0.7) +
  #geom_text(check_overlap = T, hjust = 0, nudge_x = 1.5, nudge_y = 1.5, colour = "black") +
  ggrepel::geom_text_repel(aes(label = seurat_clusters), size = 3) +
  theme_bw()
```

## Top 10 markers in each cluster
```{r, fig.height=12}
integrated <- JoinLayers(integrated)
imm.comb.markers <- FindAllMarkers(integrated, only.pos = T, min.pct = 0.25, logfc.threshold = 0.25)

imm.comb.markers %>%
    group_by(cluster) %>%
    top_n(n = 15, wt = avg_log2FC) -> top15
FeaturePlot(integrated, features = sample(top15$gene, 10), reduction = "umap_cca", ncol = 2)
```

```{r}
#DT::datatable(top15)

#write.csv2(top15, "integrated_data_noTBcells_cluster_markers_cellbender_onlyhuman.xlsx", row.names = F, quote = F)

imm.comb.markers %>%
    group_by(cluster) %>%
    top_n(n = 10, wt = avg_log2FC) -> top10

DimPlot(integrated, group.by = "seurat_clusters",reduction = "umap_cca", label = "T") + ggtitle("res=0.5")
```

```{r, fig.width=10, fig.height = 10}
DefaultAssay(integrated) <- "RNA"
DoHeatmap(integrated, features = top10$gene)
#ggsave("heatmap_integrated_s.clusters_markers_res0.5_9clusters.png", dpi = 300, width = 30, height = 25, units = "cm")

```

```{r}
#integrated <- RunAzimuth(integrated, reference = "pbmcref")
```

```{r}
DefaultAssay(integrated) <- "RNA"
# Markers

# T cell acitvation markers
# NKG7, PRF1, GNLY, and GZMA/B/H/K/M

FeaturePlot(integrated, features = c("NKG7", "PRF1", "GNLY", "GZMA", "GZMB", "GZMH", "GZMK", "GZMM"), reduction = "umap_cca")
```

```{r}
# exhaustion markers
# LAG3, TIGIT, and PDCD1
FeaturePlot(integrated, features = c("LAG3", "TIGIT", "PDCD1"), , reduction = "umap_cca")
```

```{r}
# MHC II genes
#HLA-DRB1/DPA1/DPB1/DRA/DQA1/DRB5

FeaturePlot(integrated, features = c("HLA-DRB1", "HLA-DPA1", "HLA-DPB1", "HLA-DRA", "HLA-DQA1", "HLA-DRB5"), , reduction = "umap_cca")
```

```{r}
# naive memory markers for non-expanded T cells
# CD45RA, CD62L, CD127, CD132, and CCR7
FeaturePlot(integrated, features = c("CD45RA", "CD62L", "CD127", "CD132","CCR7"), , reduction = "umap_cca")


# For CD8 T cells: https://www.sciencedirect.com/science/article/pii/S2666979X25000989
# TCF7, IL7R, CCR7, SELL, and LEF1
FeaturePlot(integrated, features = c("TCF7", "IL7R", "CCR7", "SELL","LEF1"), , reduction = "umap_cca")

```

```{r}
# TBX21
FeaturePlot(integrated, features = c("GZMB", "GZMA"), , reduction = "umap_cca")

FeaturePlot(integrated, features = c("FOXP3"), , reduction = "umap_cca")

```

```{r}
# surface protein markers
# CD137 (gene TNFRSF9), CD38, CD40L (gene CD40LG )(aka CD154), CD69, OX40 (gene TNFRSF4) (aka CD134), CD27, CD25 (gene IL2RA)
FeaturePlot(integrated, features = c("TNFRSF9", "CD38", "CD40LG", "CD38", "CD69", "TNFRSF4", "CD27", "IL2RA"), , reduction = "umap_cca")
 # CD40L was not found
```
