==> Scripts/S5_post_integration_clonotype_association.Rmd <==
---
title: "Clonotype information to integrated object"
output:
  html_document:
    df_print: paged
---


```{r opts, echo = F}
knitr::opts_chunk$set(warning = F, message = F)
# knitr::opts_chunk$set(fig.pos = "!H", out.extra = "")
#knitr::opts_chunk$set(tidy.opts=list(width.cutoff=80),tidy=TRUE)
knitr::opts_chunk$set(fig.width=10, fig.height=6) 

```


```{r load_libraries, echo = F, message = F, warning = F}
library(Seurat)
library(patchwork)
library(readxl)
library(tidyverse)
library(ggplot2)
library(ggpubr)
library(reshape2)
library(broom)
library(stringr)
library(ggplot2)
library(cowplot)
library(RColorBrewer)
library(data.table)
library(DropletUtils)
# devtools::install_github("satijalab/seurat", "seurat5")
# devtools::install_github("satijalab/seurat-data", "seurat5")
# BiocManager::install("DirichletMultinomial")
# install.packages("hdf5r")
# remotes::install_github("mojaveazure/seurat-disk")
# BiocManager::install("TFBSTools")
# remotes::install_github('satijalab/azimuth', ref = 'master')
library(Azimuth)

```


```{r}
integrated <- readRDS("robjects/naive_merged_with_integrated_proliferated_L1_L2_custom_naive_cellsnp.rds")
clonotype <- read.csv2("../Data/HUSTLEseq_mapped_clean.csv")
```

```{r}
integrated@meta.data <- integrated@meta.data %>% 
  mutate(Barcode = rownames(.)) %>% 
  mutate(Barcode = gsub(".*_", "", Barcode))

umap <- integrated@reductions[["umap_cca"]]@cell.embeddings[,1:2]

integrated_df <- data.frame(integrated@meta.data[,c("orig.ident", "Donor_expt", 
                                                    "seurat_clusters", "HTO_classification", 
                                                    "pep.pool", "donor_pep.pool", 
                                                    "Barcode")], umap)

pheno_clono <- left_join(clonotype, integrated_df, by = "Barcode")

#write.csv2(pheno_clono, "tables/phenotype_clonotype_smallset_7k.csv", row.names = F, quote = F)
```

```{r}
# T cells where TCRs were labeled as epitopes
ggplot(pheno_clono,
       aes(x = umapcca_1, y = umapcca_2, color = factor(Epitope.1))) +
  geom_point(alpha = 0.9, size = 0.6) +
  scale_color_discrete(na.value = "gray80") +
  theme_bw()
  
#ggsave("figures/umap_mapped_epitopes.png", dpi = 300)
```

```{r}
pheno_clono %>% 
  filter(Mapped == "TRUE") %>% 
ggplot(.,
       aes(x = umapcca_1, y = umapcca_2, color = factor(Clone.ID))) +
  geom_point(alpha = 0.9, size = 0.6) +
  scale_color_discrete(na.value = "gray80") +
  theme_bw()

pheno_clono %>% 
  filter(Epitope.1 == "24") %>% 
ggplot(.,
       aes(x = umapcca_1, y = umapcca_2, color = factor(Clone.ID))) +
  geom_point(alpha = 0.9, size = 0.6) +
  scale_color_discrete(na.value = "gray80") +
  facet_wrap(.~seurat_clusters) +
  theme_bw()
```


```{r}
# map clonotypes to clusters
table(pheno_clono$Epitope.1)

table(pheno_clono$Epitope.1, pheno_clono$seurat_clusters)
  
ggplot(pheno_clono,
       aes(x = umapcca_1, y = umapcca_2, color = factor(Epitope.1))) +
  geom_point(alpha = 0.7, size = 0.6) +
  scale_color_discrete(na.value = "gray80") +
  facet_wrap(. ~ Epitope.1) +
  theme_bw()
  
#ggsave("figures/umap_mapped_epitopes_to_s.clusters.png", dpi = 300)
```

```{r}
# background or bystander

table(pheno_clono$Background)


p1 <- ggplot(pheno_clono,
       aes(x = umapcca_1, y = umapcca_2, color = Background)) +
  geom_point(alpha = 0.7, size = 0.6) +
  scale_color_discrete(na.value = "gray80") +
  theme_bw()

table(pheno_clono$Bystander)


p2 <- ggplot(pheno_clono,
       aes(x = umapcca_1, y = umapcca_2, color = Bystander)) +
  geom_point(alpha = 0.7, size = 0.6) +
  scale_color_discrete(na.value = "gray80") +
  theme_bw()

p1 + p2

#ggsave("umap_background_bystander.png", dpi = 300, width = 30, height = 20, units = "cm")

table(pheno_clono$Background, pheno_clono$Epitope.1)

table(pheno_clono$Background, pheno_clono$seurat_clusters)

```
